backend/requirements.txt
Django==4.1.4
djangorestframework